Restructure hosted schema URLs
