+++ apps/site/public/types/modules/graph/0.3/schema/aggregation-operation.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/aggregation-operation",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/aggregation-operation",
   "title": "AggregationOperation",
   "type": "object",
   "properties": {
