+++ apps/site/public/types/modules/graph/0.3/schema/base-url.json
@@ -1,6 +1,6 @@
 
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/base-url",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/base-url",
   "title": "Base URL",
   "description": "The base URL of a Block Protocol ontology type (the $id of the schema, without the versioned suffix). It should a valid URL, with a trailing slash.",
   "type": "string",
