+++ apps/site/public/types/modules/graph/0.3/schema/data-type-reference.json
@@ -1,12 +1,12 @@
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/data-type-reference",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/data-type-reference",
   "title": "Data Type Reference",
   "description": "A reference to a data type.",
   "type": "object",
   "properties": {
     "$ref": {
       "$comment": "The versioned URL of a data type (the $id of the data type's schema)",
-      "$ref": "https://blockprotocol.org/types/modules/graph/versioned-url"
+      "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/versioned-url"
     }
   },
   "required": [
