+++ apps/site/public/types/modules/graph/0.3/schema/entity-type-reference.json
@@ -1,12 +1,12 @@
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/entity-type-reference",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/entity-type-reference",
   "title": "Entity Type Reference",
   "description": "A reference to a entity type.",
   "type": "object",
   "properties": {
     "$ref": {
       "$comment": "The versioned URL of a entity type (the $id of the entity type's schema)",
-      "$ref": "https://blockprotocol.org/types/modules/graph/versioned-url"
+      "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/versioned-url"
     }
   },
   "required": [
