+++ apps/site/public/types/modules/graph/0.3/schema/entity-type.json
@@ -1,19 +1,19 @@
 {
   "$schema": "http://json-schema.org/draft-07/schema#",
-  "$id": "https://blockprotocol.org/types/modules/graph/entity-type",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/entity-type",
   "title": "Entity Type",
   "description": "Specifies the structure of a Block Protocol entity type",
   "type": "object",
   "properties": {
     "schema": {
       "type": "string",
-      "const": "https://blockprotocol.org/types/modules/graph/entity-type"
+      "const": "https://blockprotocol.org/types/modules/graph/0.3/schema/entity-type"
     },
     "kind": {
       "const": "entityType"
     },
     "$id": {
-      "$ref": "https://blockprotocol.org/types/modules/graph/versioned-url"
+      "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/versioned-url"
     },
     "type": {
       "const": "object"
@@ -50,19 +50,19 @@
       }
     },
     "properties": {
-      "$ref": "https://blockprotocol.org/types/modules/graph/property-type-object"
+      "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/property-type-object"
     },
     "required": {
       "type": "array",
       "items": {
-        "$ref": "https://blockprotocol.org/types/modules/graph/base-url"
+        "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/base-url"
       }
     },
     "links": {
       "type": "object",
       "propertyNames": {
         "$comment": "Property names must be a valid versioned URL to a link entity type to allow for outgoing links of that type from this entity.",
-        "$ref": "https://blockprotocol.org/types/modules/graph/versioned-url"
+        "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/versioned-url"
       },
       "patternProperties": {
         ".*": {
@@ -118,7 +118,7 @@
         "oneOf": {
           "type": "array",
           "items": {
-            "$ref": "https://blockprotocol.org/types/modules/graph/entity-type-reference"
+            "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/entity-type-reference"
           }
         }
       },
