+++ apps/site/public/types/modules/graph/0.3/schema/entity.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/entity",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/entity",
   "title": "Entity",
   "type": "object",
   "properties": {
@@ -26,7 +26,7 @@
         },
         "entityTypeId": {
           "$comment": "The versioned URL of the entity's entity type (the $id of the entity type's schema)",
-          "$ref": "https://blockprotocol.org/types/modules/graph/versioned-url"
+          "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/versioned-url"
         }
       },
       "required": [
