+++ apps/site/public/types/modules/graph/0.3/schema/graph-resolve-depths.json
@@ -1,8 +1,8 @@
 {
-  "$id": "https://blockprotocol.org/types/modules/graph/graph-resolve-depths",
+  "$id": "https://blockprotocol.org/types/modules/graph/0.3/schema/graph-resolve-depths",
   "title": "Graph Resolve Depths",
   "description": "The fully-specified depths associated with a subgraph query.",
-  "$ref": "https://blockprotocol.org/types/modules/graph/partial-graph-resolve-depths",
+  "$ref": "https://blockprotocol.org/types/modules/graph/0.3/schema/partial-graph-resolve-depths",
   "required": [
     "hasLeftEntity",
     "hasRightEntity",
