+++ apps/site/public/types/modules/graph/0.3/schema/partial-graph-resolve-depths.json
