+++ apps/site/public/types/modules/graph/0.3/schema/property-type.json
