--- apps/site/public/types/modules/graph/link.json
@@ -1,21 +1,0 @@
-{
-  "$id": "https://blockprotocol.org/types/modules/graph/link",
-  "title": "Link",
-  "type": "object",
-  "properties": {
-    "sourceEntityId": {
-      "type": "string"
-    },
-    "destinationEntityId": {
-      "type": "string"
-    },
-    "path": {
-      "type": "string"
-    },
-    "index": {
-      "type": "integer",
-      "inclusiveMinimum": 0
-    }
-  },
-  "required": ["sourceEntityId", "destinationEntityId", "path"]
-}
