--- apps/site/public/types/modules/graph/linked-aggregation-definition.json
@@ -1,17 +1,0 @@
-{
-  "$id": "https://blockprotocol.org/types/modules/graph/linked-aggregation-definition",
-  "title": "LinkedAggregationDefinition",
-  "type": "object",
-  "properties": {
-    "sourceEntityId": {
-      "type": "string"
-    },
-    "path": {
-      "type": "string"
-    },
-    "operation": {
-      "$ref": "https://blockprotocol.org/types/modules/graph/aggregation-operation"
-    }
-  },
-  "required": ["sourceEntityId", "operation", "path"]
-}
