+++ libs/@blockprotocol/graph/src/shared/codegen/entity-type-meta-schema.json
@@ -1,5 +1,5 @@
 {
-  "$id": "https://blockprotocol.org/type-system/0.3/schema/meta/entity-type",
+  "$id": "https://blockprotocol.org/type-system/0.3/schema/entity-type",
   "title": "Entity Type",
   "description": "Specifies the structure of an Entity Type",
   "type": "object",
